Finished part4

+++ part4/bloglist/controllers/blogs.js
@@ -1,3 +1,4 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -12,8 +13,15 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+
   try {
-    const user = await User.findOne()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
       title: body.title,
@@ -23,11 +31,11 @@ blogsRouter.post('/', async (request, response, next) => {
       user: user._id
     })
 
-    const newBlog = await blog.save()
-    user.blogs = user.blogs.concat(newBlog._id)
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
 
-    response.status(201).json(newBlog.toJSON())
+    response.status(201).json(savedBlog.toJSON())
   } catch (exception) {
     next(exception)
   }
@@ -53,7 +61,22 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     response.status(204).end()
   } catch (exception) {
     next(exception)
